requirements.txt
aiogram>=3.12.0
psycopg[binary]>=3.2.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
Pillow>=10.2.0
qrcode>=7.4.2
fastapi==0.115.1
uvicorn==0.31.0
jinja2==3.1.4
